refactor: myProjectHistory 관련 컴포넌트 이동
